refactor: http header key 객체 생성
